ci: add new workflows for Teams notification and Lambda deployment
Add two new GitHub workflows: one for sending notifications to Microsoft Teams and another for building, testing, and deploying a Node.js application to AWS Lambda. The workflows are designed to be reusable and triggered by other workflows, with inputs for node version and stage. Notifications are sent on test or deployment failures.