Adds serverless site deployment workflow
Implements a GitHub Actions workflow to deploy a static site to AWS using Serverless.

The workflow is triggered by `workflow_call`, allowing it to be invoked from other workflows. It defines inputs for node version, stage, and project name.

The workflow consists of two jobs: `deploy-architecture` and `deploy-project`. The first job deploys the base architecture, and the second deploys the actual project files to an S3 bucket and invalidates the CloudFront cache.